requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
evdev==1.7.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
gradio==4.20.1
gradio_client==0.11.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openvino==2023.3.0
openvino-telemetry==2023.2.1
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
playsound==1.3.0
playsound2==1.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyMySQL==1.1.0
pynput==1.7.6
pyparsing==3.1.2
PyQt5==5.14.0
PyQt5-sip==12.13.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.1
safetensors==0.4.2
scikit-build==0.17.6
scipy==1.12.0
selenium==4.18.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
starlette==0.36.3
sympy==1.12
tdqm==0.0.1
tiktoken==0.6.0
tk==0.1.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
trio==0.24.0
trio-websocket==0.11.1
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
webdriver-manager==4.0.1
websockets==11.0.3
Werkzeug==3.0.1
whisper_mic==1.4.0
wsproto==1.2.0
zipp==3.17.0
